implement safer html fmt
I really liked the fmt based initial draft of these methods.
This first of "safe" fmt usage intercept each format interpolation.
And calls html.EscapeString. It does this reguardless of the position in
the format html and is unaware of attributes or how to escape
appropriately in those situations.

Known issue: "%q" in element tags (attribute fields ex: `<div style=%q>`
seem to not be formatted properly and the quotes are escaped.