Improve usability
AsElement and AsText make it easier to convert types without type
assertion. I wasn't sure this was important before but it does make it
easier to use.

I also did some null checks. These _should_ not occur if callers are
doing null checks but I forgot to null check a few times and I wrote
this package.

Added .Length to text so that copying the text into the Go memory space
is not required.